Adjust report tests for new structure
